// src/platform/plugins/shared/data_views/common/types.d.ts
export type RuntimePrimitiveTypes = Exclude<RuntimeType, 'composite'>;
/**
 * Runtime field definition
 * @public
 */
export interface RuntimeFieldBase {
  /**
   * Type of runtime field
   */
  type: RuntimeType;
  /**
   * Runtime field script
   */
  script?: {
    /**
     * Script source
     */
    source: string;
  };
}
/**
 * The RuntimeField that will be sent in the ES Query "runtime_mappings" object
 */
export type RuntimeFieldSpec = RuntimeFieldBase & {
  /**
   * Composite subfields
   */
  fields?: Record<
    string,
    {
      type: RuntimePrimitiveTypes;
    }
  >;
};
// ...
export type FieldSpec = DataViewFieldBase & {
  /**
   * Popularity count is used by discover
   */
  count?: number;
  /**
   * Description of field type conflicts across indices
   */
  conflictDescriptions?: Record<string, string[]>;
  /**
   * Field formatting in serialized format
   */
  format?: SerializedFieldFormat;
  /**
   * Elasticsearch field types used by backing indices
   */
  esTypes?: string[];
  /**
   * True if field is searchable
   */
  searchable: boolean;
  /**
   * True if field is aggregatable
   */
  aggregatable: boolean;
  /**
   * True if field is empty
   */
  isNull?: boolean;
  /**
   * True if field is a computed column, used in ES|QL to distinguish from index fields
   */
  isComputedColumn?: boolean;
  /**
   * True if can be read from doc values
   */
  readFromDocValues?: boolean;
  /**
   * True if field is indexed
   */
  indexed?: boolean;
  /**
   * Custom label for field, used for display in kibana
   */
  customLabel?: string;
  /**
   * Custom description for field, used for display in kibana
   */
  customDescription?: string;
  /**
   * Runtime field definition
   */
  runtimeField?: RuntimeFieldSpec;
  /**
   * list of allowed field intervals for the field
   */
  fixedInterval?: string[];
  /**
   * List of allowed timezones for the field
   */
  timeZone?: string[];
  /**
   * set to true if field is a TSDB dimension field
   */
  timeSeriesDimension?: boolean;
  /**
   * set if field is a TSDB metric field
   */
  timeSeriesMetric?: estypes.MappingTimeSeriesMetricType;
  /**
   * Whether short dots are enabled, based on uiSettings.
   */
  shortDotsEnable?: boolean;
  /**
   * Is this field in the mapping? False if a scripted or runtime field defined on the data view.
   */
  isMapped?: boolean;
  /**
   * Name of parent field for composite runtime field subfields.
   */
  parentName?: string;
  defaultFormatter?: string;
  /**
   * Indicates whether the field is a metadata field.
   */
  metadata_field?: boolean;
};